+++ src/main/java/com/dev/condominio/dto/user/UserRequest.java
@@ -26,7 +26,7 @@ public class UserRequest {
     public String email;
 
     @NotNull
-    public Integer apt;
+    public Short apt;
 
     @NotNull
     public Short bloco;
